feat(about): add link to neo league about page
