+++ apps/nova/messages/vi.json
@@ -134,7 +134,8 @@
         "title": "Cạnh tranh",
         "description": "Tham gia bảng xếp hạng và so sánh kỹ năng thiết kế prompt của bạn với người khác."
       }
-    }
+    },
+    "learn_more": "Tìm hiểu thêm"
   },
   "common": {
     "alpha": "Alpha",
